requirements.txt
# Agent Framework
google-adk[a2a]>=1.23

# Web Framework
uvicorn>=0.30.0
fastapi>=0.115.0

# Data & Analysis
pandas>=2.2.0
yfinance>=0.2.40

# LLM
google-genai>=1.0.0
google-cloud-aiplatform

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
redis>=5.0.0
psycopg2-binary

# HTTP
httpx>=0.27.0
requests>=2.31.0
beautifulsoup4

# Config
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0

# Scheduling
apscheduler>=3.10.0

# Logging
structlog>=24.0.0
